+++ DemoApp/AppDelegate.swift
@@ -1,7 +1,8 @@
 //
-// Copyright © 2021 Stream.io Inc. All rights reserved.
+// Copyright © 2022 Stream.io Inc. All rights reserved.
 //
 
+import Atlantis
 import StreamChat
 import UIKit
 
@@ -11,7 +12,8 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
         _ application: UIApplication,
         didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?
     ) -> Bool {
-        // Override point for customization after application launch.
+        // Uncomment this to be able to proxy WebSocket events from Proxyman.app
+        // Atlantis.start()
         true
     }