+++ Examples/MessengerClone/FetchImage.swift
@@ -1,9 +1,9 @@
 //
-// Copyright © 2021 Stream.io Inc. All rights reserved.
+// Copyright © 2022 Stream.io Inc. All rights reserved.
 //
 
-import SwiftUI
 import Nuke
+import SwiftUI
 
 // The MIT License (MIT)
 //
@@ -76,8 +76,8 @@ public final class FetchImage: ObservableObject, Identifiable {
         self.request = request
 
         // Try to display the regular image if it is available in memory cache
-        guard let imageURL = request.urlRequest.url else { return }
-        if let container = pipeline.cachedImage(for: imageURL) {
+        guard let imageURL = request.urlRequest?.url else { return }
+        if let container = pipeline.cache[imageURL] {
             image = container.image
             return // Nothing to do
         }