+++ Sources/StreamChat/APIClient/Endpoints/Payloads/MessagePayloads.swift
@@ -1,5 +1,5 @@
 //
-// Copyright © 2021 Stream.io Inc. All rights reserved.
+// Copyright © 2022 Stream.io Inc. All rights reserved.
 //
 
 import Foundation
@@ -279,8 +279,13 @@ struct MessageRequestBody: Encodable {
     }
 }
 
-/// An object describing the message replies JSON payload.
-struct MessageRepliesPayload: Decodable {
+/// An object describing pinned messages JSON payload.
+typealias PinnedMessagesPayload = MessageListPayload
+
+/// An object describing the message list JSON payload.
+typealias MessageRepliesPayload = MessageListPayload
+
+struct MessageListPayload: Decodable {
     let messages: [MessagePayload]
 }