+++ Sources/StreamChatUI/ChatMessageList/ChatMessage/ChatMessageLayoutOptionsResolver.swift
@@ -1,5 +1,5 @@
 //
-// Copyright © 2021 Stream.io Inc. All rights reserved.
+// Copyright © 2022 Stream.io Inc. All rights reserved.
 //
 
 import Foundation
@@ -28,11 +28,14 @@ open class ChatMessageLayoutOptionsResolver {
         with messages: AnyRandomAccessCollection<ChatMessage>,
         appearance: Appearance
     ) -> ChatMessageLayoutOptions {
-        let messageIndex = messages.index(messages.startIndex, offsetBy: indexPath.item)
-        guard let message = messages[safe: messageIndex] else {
-            indexNotFoundAssertion()
+        // Make sure the message exists. Sometimes when switching channels really fast on iPad's split view,
+        // it can happen that this method is called for old data from a previous channel.
+        guard indexPath.item < messages.count else {
             return []
         }
+
+        let messageIndex = messages.index(messages.startIndex, offsetBy: indexPath.item)
+        let message = messages[messageIndex]
 
         let isLastInSequence = isMessageLastInSequence(
             messageIndexPath: indexPath,